+++ app/jobs/send_final_subscription_expiry_alerts_job.rb
@@ -3,7 +3,7 @@ class SendFinalSubscriptionExpiryAlertsJob < ApplicationJob
 
   def perform
     Subscription.due_final_expiry_notice.each do |s|
-      SubscriptionMailer.final_expiry_warning(s.id).deliver_later
+      SubscriptionExpiryMailer.final_expiry_warning(s.id).deliver_later
     end
   end
 end
