+++ app/views/subscription_expiry_mailer/first_expiry_warning.text.haml
