+++ spec/mailers/subscription_spec.rb
@@ -20,14 +20,11 @@ RSpec.describe SubscriptionMailer, type: :mailer do
     # The hashing algorithm uses a random initialization vector to encrypt the token,
     # so is different every time, so we stub the token to be the same every time, so
     # it's clearer what we're testing when we test the unsubscribe link
-    token_attributes = subscription.token_attributes
     token = subscription.token
     allow_any_instance_of(Subscription).to receive(:token) { token }
-    allow_any_instance_of(Subscription).to receive(:token_attributes) { token_attributes }
     subscription
   end
   let(:body_lines) { mail.body.raw_source.lines }
-  let(:renew_link) { %r{http:\/\/localhost:3000\/subscriptions\/#{subscription.token_attributes}\/renew} }
 
   describe '#confirmation' do
     let(:mail) { SubscriptionMailer.confirmation(subscription.id) }
@@ -47,33 +44,6 @@ RSpec.describe SubscriptionMailer, type: :mailer do
 
     it 'has an unsubscribe link' do
       expect(body_lines[12]).to match(%r{http:\/\/localhost:3000\/subscriptions\/#{subscription.token}\/unsubscribe})
-    end
-  end
-
-  describe '#first_expiry_warning' do
-    let(:mail) { SubscriptionMailer.first_expiry_warning(subscription.id) }
-
-    it 'sends a warning email' do
-      expect(mail.subject).to eq(
-        I18n.t('job_alerts.expiry.email.first_warning.subject', reference: subscription.reference)
-      )
-      expect(mail.to).to eq([subscription.email])
-      expect(body_lines[5]).to match(/#{subscription.reference}/)
-      expect(body_lines[7]).to match(/#{I18n.t('job_alerts.expiry.email.first_warning.reset_instructions')}/)
-      expect(body_lines[9]).to match(renew_link)
-    end
-  end
-
-  describe '#final_expiry_warning' do
-    let(:mail) { SubscriptionMailer.final_expiry_warning(subscription.id) }
-
-    it 'sends a warning email' do
-      expect(mail.subject).to eq(
-        I18n.t('job_alerts.expiry.email.final_warning.subject', reference: subscription.reference)
-      )
-      expect(mail.to).to eq([subscription.email])
-      expect(body_lines[5]).to match(/#{subscription.reference}/)
-      expect(body_lines[7]).to match(renew_link)
     end
   end
 end
